requirements.txt
asgiref==3.4.1
black==21.8b0
click==8.0.1
Django==3.2.7
djangorestframework==3.12.4
flake8==3.9.2
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.3.0
psycopg2==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1
regex==2021.8.28
sqlparse==0.4.2
tomli==1.2.1
typing-extensions==3.10.0.2
